vhost: search the virtqueues driver data in read-only area
The virtqueues driver areas are read-only for the device.
While they are exposed as Read/Write to Vhost-user and
regular VDUSE backends, they are only exposed as Read-only
when the control Virtqueue is shadowed by QEMU with VDUSE
backend.

This patch makes the backend to request these areas as
Read-only, so that it can be accessed in all the
configurations.

Fixes: eefac9536a90 ("vhost: postpone device creation until rings are mapped")
Fixes: 2d1541e2b6b3 ("vhost: add vring address setup for packed queues")
Cc: stable@dpdk.org

Signed-off-by: Eugenio Pérez <eperezma@redhat.com>
Reviewed-by: Maxime Coquelin <maxime.coquelin@redhat.com>